refactor: unify logging and test file locations per AGENTS.md
Logging:
- pi-memory-mem0: console.log → console.error, fix prefix to [pi-memory-mem0]
- pi-channels, pi-computer-use: gate verbose debug logs behind DEBUG env
- pi-image-gen, pi-web-access: remove unconditional console.debug at startup

Tests:
- Move all src/*.test.ts → src/__tests__/*.test.ts for pi-memory,
  pi-memory-mem0, pi-task-scheduler, pi-attachments, pi-security,
  pi-teamwork, and storage packages
- Update relative imports accordingly
