+++ packages/pi-computer-use/src/index.ts
@@ -543,20 +543,22 @@ export default function computerUseExtension(pi: ExtensionAPI): void {
         const screenshotResult = await client!.callTool('screenshot', screenshotArgs);
         const imageContent = screenshotResult.content?.find((c) => c.type === 'image' && c.data);
 
-        console.error(
-          '[pi-computer-use analyze_screenshot] screenshot result',
-          JSON.stringify(
-            {
-              window_id: params.window_id,
-              isError: screenshotResult.isError,
-              contentTypes: screenshotResult.content?.map((c) => c.type),
-              imageDataLength: imageContent?.data?.length,
-              imageMimeType: imageContent?.mimeType,
-            },
-            null,
-            2,
-          ),
-        );
+        if (process.env.DEBUG?.includes('pi-computer-use')) {
+          console.error(
+            '[pi-computer-use analyze_screenshot] screenshot result',
+            JSON.stringify(
+              {
+                window_id: params.window_id,
+                isError: screenshotResult.isError,
+                contentTypes: screenshotResult.content?.map((c) => c.type),
+                imageDataLength: imageContent?.data?.length,
+                imageMimeType: imageContent?.mimeType,
+              },
+              null,
+              2,
+            ),
+          );
+        }
 
         if (!imageContent?.data) {
           const errorText =
@@ -582,17 +584,19 @@ export default function computerUseExtension(pi: ExtensionAPI): void {
           imageContent.mimeType ?? 'image/png',
         );
 
-        console.error(
-          '[pi-computer-use analyze_screenshot] vision analysis',
-          JSON.stringify(
-            {
-              analysisLength: analysis.length,
-              analysisPreview: analysis.slice(0, 200),
-            },
-            null,
-            2,
-          ),
-        );
+        if (process.env.DEBUG?.includes('pi-computer-use')) {
+          console.error(
+            '[pi-computer-use analyze_screenshot] vision analysis',
+            JSON.stringify(
+              {
+                analysisLength: analysis.length,
+                analysisPreview: analysis.slice(0, 200),
+              },
+              null,
+              2,
+            ),
+          );
+        }
 
         return {
           content: [{ type: 'text' as const, text: analysis }],
