+++ packages/storage/src/__tests__/session-stores.test.ts
@@ -3,7 +3,7 @@ import os from 'node:os';
 import path from 'node:path';
 import type { ConversationTurn, RuntimeSession } from '@amaster.ai/pi-shared';
 import { afterEach, describe, expect, it } from 'vitest';
-import { JsonFileConversationStore, JsonFileTranscriptStore } from './session-stores.js';
+import { JsonFileConversationStore, JsonFileTranscriptStore } from '../session-stores.js';
 
 const tmpDirs: string[] = [];
